csv_rag_backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pandas
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
uvicorn